[HOROSCOPE-20] fix compile error

+++ horoscope-core/src/main/scala/com/didichuxing/horoscope/runtime/FlowExecutorImpl.scala
@@ -45,10 +45,10 @@ class FlowExecutorImpl(
   }
 
   private val contextCache: Cache[String, Map[String, TraceVariable]] = CacheBuilder.newBuilder()
-    .maximumSize(Try(config.getInt("horoscope.flow-executor.context-cache.maximum-size")).getOrElse(100000))
+    .maximumSize(Try(config.getLong("horoscope.flow-executor.context-cache.maximum-size")).getOrElse(100000))
     .concurrencyLevel(128)
     .expireAfterAccess(
-      Try(config.getInt("horoscope.flow-executor.context-cache.expire-after-access-minutes")).getOrElse(20),
+      Try(config.getLong("horoscope.flow-executor.context-cache.expire-after-access-minutes")).getOrElse(20),
       TimeUnit.MINUTES
     ).build()
